Add support for histogram visualizations
