+++ ui/package.json
@@ -9,7 +9,8 @@
     "react": "^16.3.0",
     "react-dom": "^16.3.0",
     "react-scripts": "^1.1.4",
-    "react-select": "^2.0.0"
+    "react-select": "^2.0.0",
+    "vega": "^4.3.0"
   },
   "scripts": {
     "lint": "prettier --list-different **/*.js **/*.css",
